+++ src/test/kotlin/no/ssb/metadata/vardef/controllers/patches/DeleteTests.kt
@@ -1,0 +1,17 @@
+package no.ssb.metadata.vardef.controllers.patches
+
+import io.restassured.specification.RequestSpecification
+import no.ssb.metadata.vardef.utils.BaseVardefTest
+import no.ssb.metadata.vardef.utils.INCOME_TAX_VP1_P1
+import org.junit.jupiter.api.Test
+
+class DeleteTests : BaseVardefTest() {
+    @Test
+    fun `delete request`(spec: RequestSpecification) {
+        spec
+            .`when`()
+            .delete("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/patches")
+            .then()
+            .statusCode(405)
+    }
+}
