+++ src/test/kotlin/no/ssb/metadata/vardef/controllers/validityperiods/CompanionObject.kt
@@ -1,0 +1,40 @@
+package no.ssb.metadata.vardef.controllers.validityperiods
+
+import org.json.JSONObject
+
+class CompanionObject {
+    companion object {
+        @JvmStatic
+        fun allMandatoryFieldsChanged(): String =
+            JSONObject()
+                .apply {
+                    put("valid_from", "2024-01-11")
+                    put(
+                        "definition",
+                        JSONObject().apply {
+                            put("nb", "Intektsskatt atter ny definisjon")
+                            put("nn", "Intektsskatt atter ny definisjon")
+                            put("en", "Yet another definition")
+                        },
+                    )
+                }.toString()
+
+        @JvmStatic
+        fun noneMandatoryFieldsChanged(): String {
+            val testCase =
+                JSONObject()
+                    .apply {
+                        put("valid_from", "2021-01-01")
+                        put(
+                            "definition",
+                            JSONObject().apply {
+                                put("nb", "Intektsskatt ny definisjon")
+                                put("nn", "Intektsskatt ny definisjon")
+                                put("en", "Income tax new definition")
+                            },
+                        )
+                    }.toString()
+            return testCase
+        }
+    }
+}
