+++ src/test/kotlin/no/ssb/metadata/vardef/controllers/validityperiods/CreateTests.kt
@@ -1,57 +1,39 @@
-package no.ssb.metadata.vardef.controllers
+package no.ssb.metadata.vardef.controllers.validityperiods
 
 import io.micronaut.http.HttpStatus
 import io.restassured.http.ContentType
 import io.restassured.specification.RequestSpecification
 import no.ssb.metadata.vardef.constants.ACTIVE_GROUP
+import no.ssb.metadata.vardef.controllers.validityperiods.CompanionObject.Companion.allMandatoryFieldsChanged
+import no.ssb.metadata.vardef.controllers.validityperiods.CompanionObject.Companion.noneMandatoryFieldsChanged
 import no.ssb.metadata.vardef.models.CompleteResponse
 import no.ssb.metadata.vardef.utils.*
 import org.assertj.core.api.Assertions.assertThat
 import org.hamcrest.CoreMatchers.containsString
-import org.hamcrest.Matchers.*
+import org.hamcrest.Matchers.hasKey
 import org.json.JSONObject
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.MethodSource
 import java.time.LocalDate
 
-class ValidityPeriodsControllerTest : BaseVardefTest() {
-    /**
-     * Test input data new valid validity period
-     */
-    companion object {
-        @JvmStatic
-        fun allMandatoryFieldsChanged(): String =
-            JSONObject()
-                .apply {
-                    put("valid_from", "2024-01-11")
-                    put(
-                        "definition",
-                        JSONObject().apply {
-                            put("nb", "Intektsskatt atter ny definisjon")
-                            put("nn", "Intektsskatt atter ny definisjon")
-                            put("en", "Yet another definition")
-                        },
-                    )
-                }.toString()
-
-        @JvmStatic
-        fun noneMandatoryFieldsChanged(): String {
-            val testCase =
-                JSONObject()
-                    .apply {
-                        put("valid_from", "2021-01-01")
-                        put(
-                            "definition",
-                            JSONObject().apply {
-                                put("nb", "Intektsskatt ny definisjon")
-                                put("nn", "Intektsskatt ny definisjon")
-                                put("en", "Income tax new definition")
-                            },
-                        )
-                    }.toString()
-            return testCase
-        }
+class CreateTests : BaseVardefTest() {
+    @Test
+    fun `create new patch incorrect active group`(spec: RequestSpecification) {
+        spec
+            .given()
+            .contentType(ContentType.JSON)
+            .body(allMandatoryFieldsChanged())
+            .auth()
+            .oauth2(
+                JwtTokenHelper
+                    .jwtTokenSigned(
+                        daplaTeams = listOf("play-enhjoern-b"),
+                        daplaGroups = listOf("play-enhjoern-b-developers"),
+                    ).parsedString,
+            ).queryParam(ACTIVE_GROUP, "play-enhjoern-b-developers")
+            .`when`()
+            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
+            .then()
+            .statusCode(HttpStatus.FORBIDDEN.code)
     }
 
     @Test
@@ -65,14 +47,12 @@ class ValidityPeriodsControllerTest : BaseVardefTest() {
             .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
             .then()
             .statusCode(201)
-
         val lastPatchInSecondToLastValidityPeriod =
             validityPeriods
                 .getAsMap(INCOME_TAX_VP1_P1.definitionId)
                 .let { it.values.elementAt(it.values.size - 2) }
                 ?.last()
         val lastPatch = patches.latest(INCOME_TAX_VP1_P1.definitionId)
-
         assertThat(
             lastPatch.validUntil,
         ).isNull()
@@ -107,32 +87,6 @@ class ValidityPeriodsControllerTest : BaseVardefTest() {
     }
 
     @Test
-    fun `create new validity period missing language`(spec: RequestSpecification) {
-        val definitionNotChangedForAll =
-            JSONObject(noneMandatoryFieldsChanged())
-                .apply {
-                    put("valid_from", "2040-01-11")
-                    put(
-                        "definition",
-                        JSONObject().apply {
-                            put("nb", "Intektsskatt i økonomi")
-                            put("nn", "Intektsskatt i økonomi")
-                        },
-                    )
-                }.toString()
-
-        spec
-            .given()
-            .contentType(ContentType.JSON)
-            .body(definitionNotChangedForAll)
-            .queryParam(ACTIVE_GROUP, TEST_DEVELOPERS_GROUP)
-            .`when`()
-            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-            .then()
-            .statusCode(400)
-    }
-
-    @Test
     fun `create new validity period definition text is not changed`(spec: RequestSpecification) {
         val definitionNotChanged =
             JSONObject(noneMandatoryFieldsChanged())
@@ -153,28 +107,48 @@ class ValidityPeriodsControllerTest : BaseVardefTest() {
     }
 
     @Test
-    fun `create new validity period invalid valid from`(spec: RequestSpecification) {
-        val invalidValidFrom =
-            JSONObject(allMandatoryFieldsChanged())
-                .apply {
-                    put("valid_from", "1990-05-11")
-                }.toString()
-
+    fun `create new validity period invalid active group`(spec: RequestSpecification) {
         spec
             .given()
             .contentType(ContentType.JSON)
-            .body(invalidValidFrom)
-            .queryParam(ACTIVE_GROUP, TEST_DEVELOPERS_GROUP)
+            .body(allMandatoryFieldsChanged())
+            .queryParam(ACTIVE_GROUP, "invalid-developers")
             .`when`()
             .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
             .then()
-            .statusCode(400)
-            .body(
-                containsString(
-                    "The date selected cannot be added because it falls between previously added valid " +
-                        "from dates.",
-                ),
-            )
+            .statusCode(401)
+    }
+
+    @Test
+    fun `create new validity period no active group`(spec: RequestSpecification) {
+        spec
+            .given()
+            .contentType(ContentType.JSON)
+            .body(allMandatoryFieldsChanged())
+            .`when`()
+            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
+            .then()
+            .statusCode(403)
+    }
+
+    @Test
+    fun `create new validity period principal not owner`(spec: RequestSpecification) {
+        spec
+            .given()
+            .contentType(ContentType.JSON)
+            .auth()
+            .oauth2(
+                JwtTokenHelper
+                    .jwtTokenSigned(
+                        daplaTeams = listOf("some-other-team"),
+                        daplaGroups = listOf("some-other-team-developers"),
+                    ).parsedString,
+            ).queryParam(ACTIVE_GROUP, "some-other-team-developers")
+            .body(allMandatoryFieldsChanged())
+            .`when`()
+            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
+            .then()
+            .statusCode(403)
     }
 
     @Test
@@ -337,31 +311,54 @@ class ValidityPeriodsControllerTest : BaseVardefTest() {
     }
 
     @Test
-    fun `list validity periods`(spec: RequestSpecification) {
+    fun `create new validity period invalid valid from`(spec: RequestSpecification) {
+        val invalidValidFrom =
+            JSONObject(allMandatoryFieldsChanged())
+                .apply {
+                    put("valid_from", "1990-05-11")
+                }.toString()
+
         spec
+            .given()
+            .contentType(ContentType.JSON)
+            .body(invalidValidFrom)
+            .queryParam(ACTIVE_GROUP, TEST_DEVELOPERS_GROUP)
             .`when`()
-            .get("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
+            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
             .then()
-            .statusCode(200)
-            .body("size()", `is`(2))
-            .body("[0].valid_from", equalTo("1980-01-01"))
-            .body("[0].patch_id", equalTo(7))
-            .body("[1].valid_from", equalTo("2021-01-01"))
-            .body("[1].patch_id", equalTo(6))
+            .statusCode(400)
+            .body(
+                containsString(
+                    "The date selected cannot be added because it falls between previously added valid " +
+                        "from dates.",
+                ),
+            )
     }
 
     @Test
-    fun `list validity periods return type`(spec: RequestSpecification) {
-        val body =
-            spec
-                .`when`()
-                .get("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-                .then()
-                .statusCode(HttpStatus.OK.code)
-                .extract()
-                .body()
-                .asString()
-        assertThat(jsonMapper.readValue(body, Array<CompleteResponse>::class.java)).isNotNull
+    fun `create new validity period missing language`(spec: RequestSpecification) {
+        val definitionNotChangedForAll =
+            JSONObject(noneMandatoryFieldsChanged())
+                .apply {
+                    put("valid_from", "2040-01-11")
+                    put(
+                        "definition",
+                        JSONObject().apply {
+                            put("nb", "Intektsskatt i økonomi")
+                            put("nn", "Intektsskatt i økonomi")
+                        },
+                    )
+                }.toString()
+
+        spec
+            .given()
+            .contentType(ContentType.JSON)
+            .body(definitionNotChangedForAll)
+            .queryParam(ACTIVE_GROUP, TEST_DEVELOPERS_GROUP)
+            .`when`()
+            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
+            .then()
+            .statusCode(400)
     }
 
     @Test
@@ -382,97 +379,5 @@ class ValidityPeriodsControllerTest : BaseVardefTest() {
 
         val completeResponse = jsonMapper.readValue(body, CompleteResponse::class.java)
         assertThat(completeResponse).isNotNull
-    }
-
-    @Test
-    fun `create new validity period invalid active group`(spec: RequestSpecification) {
-        spec
-            .given()
-            .contentType(ContentType.JSON)
-            .body(allMandatoryFieldsChanged())
-            .queryParam(ACTIVE_GROUP, "invalid-developers")
-            .`when`()
-            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-            .then()
-            .statusCode(401)
-    }
-
-    @Test
-    fun `create new validity period no active group`(spec: RequestSpecification) {
-        spec
-            .given()
-            .contentType(ContentType.JSON)
-            .body(allMandatoryFieldsChanged())
-            .`when`()
-            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-            .then()
-            .statusCode(403)
-    }
-
-    @Test
-    fun `create new validity period principal not owner`(spec: RequestSpecification) {
-        spec
-            .given()
-            .contentType(ContentType.JSON)
-            .auth()
-            .oauth2(
-                JwtTokenHelper
-                    .jwtTokenSigned(
-                        daplaTeams = listOf("some-other-team"),
-                        daplaGroups = listOf("some-other-team-developers"),
-                    ).parsedString,
-            ).queryParam(ACTIVE_GROUP, "some-other-team-developers")
-            .body(allMandatoryFieldsChanged())
-            .`when`()
-            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-            .then()
-            .statusCode(403)
-    }
-
-    @Test
-    fun `create new patch incorrect active group`(spec: RequestSpecification) {
-        spec
-            .given()
-            .contentType(ContentType.JSON)
-            .body(allMandatoryFieldsChanged())
-            .auth()
-            .oauth2(
-                JwtTokenHelper
-                    .jwtTokenSigned(
-                        daplaTeams = listOf("play-enhjoern-b"),
-                        daplaGroups = listOf("play-enhjoern-b-developers"),
-                    ).parsedString,
-            ).queryParam(ACTIVE_GROUP, "play-enhjoern-b-developers")
-            .`when`()
-            .post("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-            .then()
-            .statusCode(HttpStatus.FORBIDDEN.code)
-    }
-
-    @Test
-    fun `list validity periods unauthenticated`(spec: RequestSpecification) {
-        spec
-            .given()
-            .auth()
-            .none()
-            .`when`()
-            .get("/variable-definitions/${INCOME_TAX_VP1_P1.definitionId}/validity-periods")
-            .then()
-            .statusCode(HttpStatus.UNAUTHORIZED.code)
-    }
-
-    @ParameterizedTest
-    @MethodSource("no.ssb.metadata.vardef.utils.TestUtils#definitionIdsAllStatuses")
-    fun `list validity periods authenticated`(
-        definitionId: String,
-        expectedStatus: String,
-        spec: RequestSpecification,
-    ) {
-        spec
-            .`when`()
-            .get("/variable-definitions/$definitionId/validity-periods")
-            .then()
-            .statusCode(HttpStatus.OK.code)
-            .body("[0].variable_status", equalTo(expectedStatus))
     }
 }
